Add files via upload
Investment Bankers . CA's . Hedge Fund / Portfolio Managers . Forex traders . Commodities Analysts. These have been historically considered to be among the most coveted professions of all time. Yet, if one fails to keep up with the demands of the day, one would find one's skills to be obsolete in this era of data analysis. Data Science has inarguable been the hottest domain of the decade, asserting its need in every single sphere of corporate life. It was not long ago when we discovered the massive potential of incorporating ML/AI in the financial world. Now, the very idea of the two being disjointed sounds strange. Data Science has been incremental in providing powerful insights ( which people didn't even know existed ) and helped massively increase the efficiency, helping everyone from a scalp trader to a long term debt investor. Accurate predictions, unbiased analysis, powerful tools that run through millions of rows of data in the blink of an eye have transformed the industry in ways we could've never imagined. The following program is designed to both test your knowledge and to give you the feel and experience of a real world financial world - data science problem.

Steps to complete this project:-
Go through the "Basics of Financial Market" pdf to understand the basic terminologies of stock market.
Go through the instructions in the respective modules to understand the tasks assigned for each module
Go through the format notebooks for writing the solutions for the respective modules in the correct format.
Edit the solution jupyter notebooks and add your code for the queries in the respective modules or uplaod your notebook for that module.
Note: Only .ipynb files are supported.Other modules will be uploaded after I get the solutions for the current modules.
Disclaimer before contributing: Only significant contributions to this project would be accepted.